+++ server/src/app.module.ts
@@ -7,17 +7,9 @@ import { ConfigModule } from "@nestjs/config";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 import { RedisModule } from "./modules/redis/redis.module";
-import redisConfig from "./config/redis.config";
 
 @Module({
-  imports: [
-    ConfigModule.forRoot({ isGlobal: true }),
-    AuthModule,
-    GameModule,
-    PrismaModule,
-    RedisModule.registerAsync(redisConfig),
-    ChatModule,
-  ],
+  imports: [ConfigModule.forRoot({ isGlobal: true }), AuthModule, GameModule, PrismaModule, RedisModule, ChatModule],
 
   controllers: [AppController],
