+++ server/src/modules/game/game.module.ts
@@ -5,10 +5,9 @@ import { GameController } from "./game.controller";
 import { GameService } from "./game.service";
 import { PrismaModule } from "src/modules/prisma/prisma.module";
 import { RedisModule } from "../redis/redis.module";
-import redisConfig from "src/config/redis.config";
 
 @Module({
-  imports: [PrismaModule, RedisModule.registerAsync(redisConfig)],
+  imports: [PrismaModule, RedisModule],
   controllers: [GameController],
   providers: [GameService],
 })
