+++ server/src/modules/jwt/jwt.module.ts
@@ -1,8 +1,9 @@
 import { ConfigService } from "@nestjs/config";
+import { JwtModule as JwtModuleWithoutRegister } from "@nestjs/jwt";
 
-export default {
+export const JwtModule = JwtModuleWithoutRegister.registerAsync({
   useFactory: async (config: ConfigService) => ({
     secret: config.get<string>("JWT_SECRET_KEY"),
   }),
   inject: [ConfigService],
-};
+});
